requirements.txt
numpy
pyvcf>=0.6.7
biopython==1.68
bx-python==0.7.1
cython
pysam
hgvs==0.3.7
pytest
pytest-cov
six